backend/requirements.txt
tornado==4.4.2
greenlet==0.4.10
uwsgi==2.0.14
SQLAlchemy==1.1.4
psycopg2==2.6.2
pytz==2016.7
python-dateutil==2.6.0
